requirements.txt
flask
flask_cors
flask_migrate
pillow
ultralytics
shapely
pyclipper
scikit-image
tensorflow
onnxruntime

#YOLOのtfliteモデル変換時に必要なパッケージ
#onnx2tf==1.22.3
#tf_keras
#onnx==1.16.1
#onnx_graphsurgeon
#sng4onnx
#onnxslim==0.1.34
#tflite-support==0.4.4